pull out irv algo, copy and modify irv implementation with lfws modifications
